update db schema and completed the assciation mapping

+++ backend/database-service/src/main/java/com/elms/databaseservice/repos/StudentRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.elms.databaseservice.models.Student;
 
-public interface StudentRepo extends JpaRepository<Student, Long> {
+public interface StudentRepo extends JpaRepository<Student, Integer> {
 	public Student findByStudentId(int id);
 }
